+++ src/apiServices/user/interfaces/users.interface.ts
