+++ src/apiServices/user/users.service.ts
@@ -1,6 +1,6 @@
 import { HttpException } from '@exceptions/HttpException';
-import { User } from '@/apiServices/user/users.interface';
-import userModel from '@/apiServices/user/users.model';
+import { User } from '@apiServices/user/interfaces/users.interface';
+import userModel from '@apiServices/user/models/users.model';
 import { isEmpty } from '@utils/util';
 
 class UserService {
