--- src/databases/index.ts
@@ -1,10 +1,0 @@
-import { DB_URL } from '@config';
-
-export const dbConnection = {
-  url: DB_URL,
-  options: {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-    useFindAndModify: false,
-  },
-};
