+++ src/server.ts
@@ -1,5 +1,5 @@
 import App from '@/app';
-import AuthController from '@/apiServices/auth/auth.controller';
+import AuthController from '@apiServices/auth/auth.controller';
 import validateEnv from '@utils/validateEnv';
 import 'reflect-metadata';
