+++ src/utils/getAuthorizationtoken.ts
@@ -1,5 +1,5 @@
-import { RequestWithUser } from '@/apiServices/auth/auth.interface';
+import { RequestWithUser } from '@/apiServices/auth/interfaces/auth.interface';
 
-export const getAuthorizationtoken = (req: RequestWithUser): String | null => {
+export const getAuthorizationToken = (req: RequestWithUser): String | null => {
   return req.cookies['Authorization'] || (req.header('Authorization') ? req.header('Authorization').split('Bearer ')[1] : null);
 };
