+++ tsconfig.json
@@ -22,6 +22,7 @@
     "baseUrl": "src",
     "paths": {
       "@/*": ["*"],
+      "@apiServices" : ["apiServices/*"],
       "@config": ["config"],
       "@controllers/*": ["controllers/*"],
       "@databases": ["databases"],
